Unifying names of msgpack structs.
This seems to change a lot of files and one hopes does not lead to merge conflicts! However, this is needed to simplify the process of creating meta-data to insert into the array headers. Another round of refactor is still needed to actually use the mpack wrapper code in each of the Apps.