Experimenting now with utilizing the exact linear solve in Wang et al. 2020 in Appendix D (the same function that is utilized in sources_implicit.c).
For now this is slightly hacky in that I have copied the functionality from sources_implicit.c into the pkpm_kernels header file as just a static function (with a few small adjustments). If this ends up working, I will propose a design review in which some functionality in sources_implicit is split into a private header that is includable by multiple functions (similar to the private headers for wv_eqn that are utilized by DG fluids and PKPM).