Adding a new option to PKPM: in 1x, can solve the PKPM system in field-aligned coordinates with the one spatial dimension being the length along the field line. In 1x, due to the nature of PKPM, this requires only a few things: 1. The user to specify the Jacobian = 1/B 2. The user to specify the corresponding div(b) = -1/B dB/dz for the mirror force the particles experience in this external field 3. The RHS of the fluid momentum equation be adjusted to Jrho*Eparallel + Jp_perp div(b) (which compared to original 1x calculations if we store Jrho instead of rho, easy enough when solving for J*f instead of f, and assume Eparallel is the 0th component of the em array, then we only have to add this second force which changes the momentum due to the local Jp_perp and externally specified div(b)) 4. The "PKPM acceleration variables" bb : grad(u) = du_parallel/dz and bb : grad(u) - div(u) = u_parallel div(b) 5. Dividing out the Jacobian from the momentum Jrho*u_parallel before incrementing to the parallel electric field (since in 1x the field equation is simply dE_parallel/dt = -Jparallel/epsilon0)
Thus with just two well-placed flags for the field-line following coordinate system for computing the source terms and acceleration variables, along with a division of the Jacobian prior to the current accumulation, 1x field-line following simulations with the PKPM system can be performed.

I'm not currently happy with the names since these names are different from other flags we utilize for general geometry, but because this is meant partially as a proof of principal that this can be done I decided to make the naming convention very distinct to start with. It would be nice to think about how this approach would generalize or if this is something wholly distinct from what we want to do going forward. For example, it would be relatively straightforward to get the desired Jacobian and div(b) from gk_geometry (whether it's an eqdsk or a mapping function). However, a significant chunk of the simplifications currently present rely on one of the momentum equations being for uparallel (not a global Cartesian basis like ux, uy, uz). At first glance the typical upside of a global Cartesian basis is not obvious in PKPM, because the momentum equation gives a separate kinetic energy equation from the internal energy evolved by PKPM, so if you have to add your different equations anyway for a total energy equation, perhaps it is worth changing the momentum solve to a (uparallel, u_perp1, u_perp2) but then we would have to think carefully about how PKPM currently utilizes the rotations in wave_geom to do the Roe solve in 2D and 3D.

Still initial mirror calculations using a Lorentzian background B field are promising, with the evolution proceeding as expected on short (microsecond) time scales in terms of the trapping of particles by the mirror force and sheath formation in the expander. Right now I only support explicit time stepping, so a small cfl_frac is chosen on coarse resolutions (the cfl can be increased if one runs at higher resolution) to resolve the electron plasma frequency at the specified parameters (which sets about a few picosecond time step).