Since we are about to do a major release of Gkeyll and I have had a couple of folks ping me about running certain simulations that can be found on Ammar's website, especially Ammar's Hasegawa-Wakatani note (https://ammar-hakim.org/sj/je/je17/je17-hasegawa-wakatani.html), I decided to utilize the canonical PB infrastructure to add the option to solve fluid equations such as incompressible Euler, Hasegawa-Mima, and Hasegawa-Wakatani to gkylzero (and I think it would be good to go a little further and add things like reduced MHD too).
The way that this works is there is now a small wv_eqn object, wv_canonical_pb_fluid, which exists solely for a user to distinguish which DG fluid system they want and to pass some equation specific parameters (such as alpha for the adiabatic coupling coefficient in Hasegawa-Wakatani).

For now, the solution of these systems is entirely internal to the vm_fluid_species.c component of the app, so if the system detects you are doing something it knows is a fluid system that can be written as a canonical Poisson bracket, vm_fluid_species internally creates an FEM Poisson solver that gives the potential (the "Hamiltonian" in these systems that defines the evolution via the Poisson brakcet) and then utilizes that quantity to get things like the surface characteristic velocity {z, phi} and perform the necessary source updates if they are present (such as alpha*(phi - n) in Hasegawa-Wakatani).

Incompressible Euler works great. Produces visually identical solutions to the simulations on Ammar's website. Still need to add the option to do Hasegawa-Wakatani by having a "two-fluid" version of the PB that updates both fluids simultaneously. I think I am initializing Hasegawa-Mima incorrectly, as the background density gradient is having periodic boundary conditions applied to it and thus all the turbulence is excited at the edge of the domain instead of in the initial Gaussian density perturbation in the center of the box. Will experiment further.