Figured out the problem with Hasegawa-Mima: I was not correctly handling the background density gradient, which in Ammar's old simulation journal is directly folded into the update of the Poisson bracket by modifying the input density before entering the updater.
To permit more flexible density gradients, I instead modify the source term based on the literature to just perform the Poisson bracket of {phi, background_n_gradient} in the source update.

Right now hard-coding the same density gradient Ammar used in his simulation journal. Will work on making this a function pointer which can be passed directly to the app and getting rid of the kappa parameter (since we are doing the full Poisson bracket of phi with the gradient now and there is no need to specify any further parameters for this coupling).