The FPO branch is encountering a truly agonizingly annoying GPU bug where the Nvidia compiler is just straight up giving up on the size of the Gkeyll system at this point. Basically we get errors that nvcc can no longer handle the optimizations it does to create temporary variables and put things in registers.
This development is unfortunate, as it has nothing to do with FPO or main per se. Both independently compile with GPUs, but their combination does not. Further, @jmrodman and I aggressively refactored parts of the FPO over the last 72 hours to make the compiler have an easier time and now the FPO is not even the worst culprit in the Gkeyll infrastructure for long lines and non-trivial compiler optimizations. Unfortunately the error persists.

As a truly desperate measure to try to solve this issue, I have gone through the code base and looked for where we have been doing things historically the GPU compiler struggles to optimize, such as other kernels which have long lines but the GPU was fine with, in the hope that the problem is not the FPO, but again, the size of Gkeyll at this point and thus try to make the whole Gkeyll system easier to compile.

1. Getting rid of remaining 3x2v_p2 gyrokinetic kernels that are not being utilized.
2. Getting rid of the vlasov_gen_geo kernels which are being superseded by canonical PB
3. Redoing the Vlasov and SR Vlasov volume kernels to eliminate some temporary variables and perform the volume updates *sequentially*, so that in 2x3v and 3x3v, we do not try to do one final super update, but a sequence of updates (which breaks up lines that were >20k large into 3 7k long lines).

May God have mercy on me and grant me salvation from my sins.