Restructuring of the canonical PB fluid systems to be a little more general and a little bit better organized. Originally I had coded up the solvers exactly like Ammar's simulation journal, but the general form of the equations does not have a turbulence drive of the form -kappa*dphi/dy, where kappa is a constant, it has a turbulence drive of the form {phi, ln(n0/wci)} where {., .} is the canonical Poisson bracket.
Since ln(n0/wci) is some general function of x and y, I have rewritten the system to take the Poisson bracket of phi with some general specified "n0," a background density. The Lua input files now specify a function called n0 that is used to set how the turbulence is driven, and there is no assumption that this turbulence drive varies only in x (though since y is the zonal direction, I am not sure if in general people do simulations with an n0 that varies in y, but the literature writes this equation as a general Poisson bracket for the turbulence drive so that's what I have coded up).

I have renamed the "background_n_gradient" variable everywhere to be either n0 or can_pb_n0 (in places where n0 is not descriptive enough). I have also reordered the inputs in a few places just because the Poisson bracket is {phi, n0} and I didn't like that the functions were taking n0 before phi as inputs in this case. Slightly OCD, but I do think this looks better.

I have also added the applied advection function pointers to the Lua wrappers and cleaned up other sides of the Vlasov public API. If I can show that the applied advection is working, I think I can declare victory on this branch and file the PR.