Fixing the advection equation in the fluid app, along with the myriad of advection and diffusion tests.
First problem was, amusingly enough, the type in wv_advect was just wrong, which probably explained some of the weird seg faulting that we were seeing and seemed slightly unintuitive for why the simulations were going into certain parts of the program flow (only made easier to debug now that there are completely separate initialization routines for advection and I could tell that it just wasn't initializing the right part of the app.

Also add the wv_advect equation to all the diffusion tests so the diffusion tests work again. I should redo the big convergence studies that originally motivated all these tests, but at least the ones I've tried run again.

Finally, hook in the wv_advect to the Lua wrappers and add the corresponding advection tests in 1x and 2x in Lua. Things are looking quite good. Great success.