Removing the warnings generated by the basis function refactor, which were caused by the fact that for some reason we were passing the basis objects by reference to the conf-phase weak multiply methods, but not making them const.
This exercise caused me to realize the basis objects are being passed by value to all the other bin_op methods *except* the conf-phase weak multiply methods and I cannot even fathom how this inconsistency arose. We should revisit this, but changing all the bin_op methods to take the basis by reference is going to be a large refactor of the app to make sure we don't mess anything up.