More regression testing cleanup, and adding equivalent Vlasov tests over a short time scale (since the test is a 1x3v one). The short time scale works anyway for the resonant case where by 10/omegaC the distribution function is encroaching on the velocity space wall.
I have added a note to the regression files that these initial conditions are a slight reconfiguration of the original tests in @ammarhakim's simulation journal (https://ammar-hakim.org/sj/je/je32/je32-vlasov-test-ptcl.html) which were originally 1x2v (constant z magnetic field with oscillating x electric field). I forgot I rotated the coordinate system for testing with PKPM (so an x magnetic field with z electric field) so that the vparallel direction was in the plane. At this point I like this test better anyway because it's such a stringent test of three of the four major apps, with identical conditions, and thus allows folks to easily look at each of the apps individually for the same input conditions.