Adding the ability to include both a gravity solve and an electrostatic potential solve within the same simulation for use in a collaboration to examine the potential impact of small q/m on gravitational instabilities (a "dark" plasma since we know if dark matter does have a charge, it is at least a million times smaller than the protons charge).
Right now this is somewhat hacky as it is controlled entirely by the parameter alpha_g, which is the ratio of the strength of the gravitational force to the forces in electromagnetism (for a real proton this number is ~10^-36). Basically if a user asks for alpha_g the app understands that to mean that you also want gravity (or gravity by itself) so it computes a different potential solve with the correct sign, and scales the RHS of the Poisson solve by alpha_g (which means that you can run purely self-gravity now by setting q=0 and alpha_g = 1.0).

Committed a "dark plasma" two-stream instability which goes Jeans unstable, but in a fun new way because it is first two-stream unstable before warping in phase space due to the gravitational interactions of the "dark electrons" and "dark positrons."