The implementation of resistivity had some limitations and hard-coding for specific transport problems we were doing with Nic when he visited last year.
I am extending resistivity in a couple of key ways, at least focusing on explicit resistivity for now:
1. Use the same tau_ei from the non_ideal private header to avoid confusing differences in definitions depending on how one defines certain constants.
2. Add the explicit temperature dependence of tau_ei by computing the model-dependent pressure prior to the update, so that we can use the local T_elc in each cell for the update.
3. For Euler, add the explicit temperature relaxation (only works with explicit time stepping right now).
4. Utilize the same coll_fac parameter that Braginskii utilizes so that one can set the collision time to be a ratio to whatever the normalized time scale is for the problem (instead of exclusively working for real SI units).

Some further clean up is desired but now the friction tests are behaving reasonably and a resistive GEM problem runs stably, with a seeming narrowing of the current sheet to a fixed length scale based on the size of the resistivity. More testing is needed but hopefully I can use this to empirically show some Sweet-Parker relations to the plasma physics class I am guest lecturing for.