Added test to check the mirror grid generator with an analytical psi = 1/2*r^2*(z^2+1). Note this is just a cooked up flux function but as our cubics are exact for this, we can get basically machine-precision metrics. The mirror grid gen is now tested agagainst this analytical geometry and tests pass.
During this process discovered a very aerious confusion with the way the cubic derivatives are computed: for some reason Akash (perhap due to miscommunication from me) returned the derivatives in the local space of the basis functions [-1,1] and not the acutal physical space. This is a cause for confusion. I have gone and fixed this and I ran the efit test. But I am not sure everything is fixed and tested as we do not have a unit test for all the various gradient functions. We need to add them! Will file an issue and have Akash add tests and fix.